Move create_storage from pyanaconda.storage
Move the function create_storage to the Storage module.
