+++ pyanaconda/modules/storage/devicetree/model.py
@@ -17,9 +17,6 @@
 #
 # Red Hat Author(s): David Lehman <dlehman@redhat.com>
 #
-
-"""This module provides storage functions related to OS installation."""
-
 import os
 
 from blivet.blivet import Blivet
@@ -42,6 +39,16 @@ from pyanaconda.modules.common.constants.services import NETWORK
 import logging
 log = logging.getLogger("anaconda.storage")
 
+__all__ = ["create_storage"]
+
+
+def create_storage():
+    """Create the storage object.
+
+    :return: an instance of the Blivet's storage object
+    """
+    return InstallerStorage()
+
 
 class InstallerStorage(Blivet):
     """ Top-level class for managing installer-related storage configuration. """
@@ -55,6 +62,12 @@ class InstallerStorage(Blivet):
         self.fsset = FSSet(self.devicetree)
         self._short_product_name = shortProductName
         self._default_luks_version = DEFAULT_LUKS_VERSION
+
+        # Set the default filesystem type.
+        self.set_default_fstype(conf.storage.file_system_type or self.default_fstype)
+
+        # Set the default LUKS version.
+        self.set_default_luks_version(conf.storage.luks_version or self.default_luks_version)
 
     @property
     def bootloader(self):
