+++ tests/nosetests/pyanaconda_tests/clearpart_test.py
@@ -4,7 +4,7 @@ import unittest.mock as mock
 import blivet
 
 from pyanaconda.modules.storage.disk_initialization import DiskInitializationConfig
-from pyanaconda.storage.initialization import create_storage
+from pyanaconda.modules.storage.devicetree import create_storage
 from pyanaconda.core.constants import CLEAR_PARTITIONS_ALL, CLEAR_PARTITIONS_LINUX, CLEAR_PARTITIONS_NONE
 from parted import PARTITION_NORMAL
 from blivet.flags import flags
